@@ -5,14 +5,13 @@ homepage: https://github.com/lukasgit/flutter_contacts
 maintainer: Lukas Dickie (@lukasgit)
 
 environment:
-  sdk: '>=2.12.0-0 <3.0.0'
-  flutter: '>=1.12.0'
+  sdk: '>=3.2.0'
 
 dependencies:
   flutter:
     sdk: flutter
-  collection: ^1.15.0
-  quiver: ^3.0.0
+  collection: ^1.19.0
+  quiver: ^3.2.2
 
 flutter:
   plugin:
