@@ -11,7 +11,7 @@ class ContactListPage extends StatefulWidget {
 }
 
 class _ContactListPageState extends State<ContactListPage> {
-  List<Contact> _contacts;
+  late List<Contact> _contacts;
 
   @override
   void initState() {
@@ -40,7 +40,7 @@ class _ContactListPageState extends State<ContactListPage> {
 
   void updateContact() async {
     Contact ninja = _contacts
-        .firstWhere((contact) => contact.familyName.startsWith("Ninja"));
+        .firstWhere((contact) => contact.familyName?.startsWith("Ninja") ?? false);
     ninja.avatar = null;
     await ContactsService.updateContact(ninja);
 
@@ -88,9 +88,9 @@ class _ContactListPageState extends State<ContactListPage> {
       body: SafeArea(
         child: _contacts != null
             ? ListView.builder(
-                itemCount: _contacts?.length ?? 0,
+                itemCount: _contacts.length ?? 0,
                 itemBuilder: (BuildContext context, int index) {
-                  Contact c = _contacts?.elementAt(index);
+                  Contact? c = _contacts.elementAt(index);
                   return ListTile(
                     onTap: () {
                       Navigator.of(context).push(MaterialPageRoute(
@@ -100,8 +100,8 @@ class _ContactListPageState extends State<ContactListPage> {
                                     contactOnDeviceHasBeenUpdated,
                               )));
                     },
-                    leading: (c.avatar != null && c.avatar.length > 0)
-                        ? CircleAvatar(backgroundImage: MemoryImage(c.avatar))
+                    leading: (c.avatar != null && (c.avatar?.isNotEmpty ?? false))
+                        ? CircleAvatar(backgroundImage: MemoryImage(c.avatar!))
                         : CircleAvatar(child: Text(c.initials())),
                     title: Text(c.displayName ?? ""),
                   );
@@ -123,7 +123,7 @@ class _ContactListPageState extends State<ContactListPage> {
 }
 
 class ContactDetailsPage extends StatelessWidget {
-  ContactDetailsPage(this._contact, {this.onContactDeviceSave});
+  ContactDetailsPage(this._contact, {required this.onContactDeviceSave});
 
   final Contact _contact;
   final Function(Contact) onContactDeviceSave;
@@ -202,7 +202,7 @@ class ContactDetailsPage extends StatelessWidget {
             ListTile(
               title: Text("Birthday"),
               trailing: Text(_contact.birthday != null
-                  ? DateFormat('dd-MM-yyyy').format(_contact.birthday)
+                  ? DateFormat('dd-MM-yyyy').format(_contact.birthday!)
                   : ""),
             ),
             ListTile(
@@ -219,9 +219,9 @@ class ContactDetailsPage extends StatelessWidget {
                   ? _contact.androidAccountType.toString()
                   : ""),
             ),
-            AddressesTile(_contact.postalAddresses),
-            ItemsTile("Phones", _contact.phones),
-            ItemsTile("Emails", _contact.emails)
+            AddressesTile(_contact.postalAddresses ?? []),
+            ItemsTile("Phones", _contact.phones ?? []),
+            ItemsTile("Emails", _contact.emails ?? [])
           ],
         ),
       ),
@@ -323,7 +323,7 @@ class _AddContactPageState extends State<AddContactPage> {
         actions: <Widget>[
           TextButton(
             onPressed: () {
-              _formKey.currentState.save();
+              _formKey.currentState?.save();
               contact.postalAddresses = [address];
               ContactsService.addContact(contact);
               Navigator.of(context).pop();
@@ -407,7 +407,7 @@ class _AddContactPageState extends State<AddContactPage> {
 }
 
 class UpdateContactsPage extends StatefulWidget {
-  UpdateContactsPage({@required this.contact});
+  UpdateContactsPage({required this.contact});
 
   final Contact contact;
 
@@ -416,7 +416,7 @@ class UpdateContactsPage extends StatefulWidget {
 }
 
 class _UpdateContactsPageState extends State<UpdateContactsPage> {
-  Contact contact;
+  late Contact contact;
   PostalAddress address = PostalAddress(label: "Home");
   final GlobalKey<FormState> _formKey = GlobalKey<FormState>();
 
@@ -438,7 +438,7 @@ class _UpdateContactsPageState extends State<UpdateContactsPage> {
               color: Colors.white,
             ),
             onPressed: () async {
-              _formKey.currentState.save();
+              _formKey.currentState?.save();
               contact.postalAddresses = [address];
               await ContactsService.updateContact(contact);
               Navigator.of(context).pushReplacement(
